requirements.txt
# package               version
numpy                   ==1.21.2
Keras                   ==2.4.3
docopt                  ==0.6.2
pandas                  ==1.3.3
Pillow                  ==8.3.2
PyInquirer              ==1.0.3
prompt-toolkit          ==1.0.14
opencv-python           ==4.5.3.56
natsort                 ==7.1.1
scikit-learn            ~=0.24.2